+++ src/prefect/utilities/executors.py
@@ -1,10 +1,12 @@
 import datetime
 import signal
+import sys
 import threading
 import time
 from concurrent.futures import ThreadPoolExecutor
 from concurrent.futures import TimeoutError as FutureTimeout
 from functools import wraps
+from logging import Logger
 from typing import TYPE_CHECKING, Any, Callable, Dict, List, Union
 
 import dask
@@ -31,13 +33,14 @@ class Heartbeat:
             to not require arguments
     """
 
-    def __init__(self, interval: int, function: Callable) -> None:
+    def __init__(self, interval: int, function: Callable, logger: Logger) -> None:
         self.interval = interval
         self.rate = min(interval, 1)
         self.function = function
+        self.logger = logger
         self._exit = False
 
-    def start(self) -> None:
+    def start(self, name_prefix: str = "") -> None:
         """
         Calling this method initiates the function calls in the background.
         """
@@ -57,17 +60,42 @@ class Heartbeat:
                 iters = (iters + 1) % self.interval
                 time.sleep(self.rate)
 
-        self.executor = ThreadPoolExecutor(max_workers=2)
-        self.fut = self.executor.submit(looper)
+        def monitor() -> None:
+            while not self._exit:
+                if not self.fut.running():
+                    self.logger.warning(
+                        "Heartbeat thread appears to have died.  This could result in a zombie run."
+                    )
+                    return
+                time.sleep(self.rate / 2)
 
-    def cancel(self) -> None:
+        kwargs = dict(max_workers=2)  # type: Dict[str, Any]
+        if sys.version_info.minor != 5:
+            kwargs.update(dict(thread_name_prefix=name_prefix))
+        self.executor = ThreadPoolExecutor(**kwargs)
+        self.fut = self.executor.submit(looper)
+        self.executor.submit(monitor)
+
+    def cancel(self) -> bool:
         """
         Calling this method after `start()` has been called will cleanup
         the background thread and cease calling the function.
+
+        Returns:
+            - bool: indicating whether the background thread was still
+                running at time of cancellation
         """
-        self._exit = True
+        running = True
         if hasattr(self, "executor"):
+            # there is a possible race condition wherein a FlowRunner can
+            # raise a KeyboardInterrupt from its heartbeat thread the exact
+            # moment after the executor is created for a TaskRunner but before
+            # the future is actually submitted to the executor
+            if hasattr(self, "fut"):
+                running = self.fut.running()
+            self._exit = True
             self.executor.shutdown()
+        return running
 
 
 def run_with_heartbeat(
@@ -82,18 +110,39 @@ def run_with_heartbeat(
     def inner(
         self: "prefect.engine.runner.Runner", *args: Any, **kwargs: Any
     ) -> "prefect.engine.state.State":
-        timer = Heartbeat(prefect.config.cloud.heartbeat_interval, self._heartbeat)
+        timer = Heartbeat(
+            prefect.config.cloud.heartbeat_interval, self._heartbeat, self.logger
+        )
+        obj = getattr(self, "task", None) or getattr(self, "flow", None)
+        thread_name = "PrefectHeartbeat-{}".format(getattr(obj, "name", "unknown"))
+        heartbeat_threads = []
         try:
             try:
                 if self._heartbeat():
-                    timer.start()
+                    timer.start(name_prefix=thread_name)
+                    heartbeat_threads.extend(
+                        [
+                            t
+                            for t in threading.enumerate()
+                            if t.name.startswith(thread_name)
+                        ]
+                    )
+                    if not heartbeat_threads and sys.version_info.minor != 5:
+                        self.logger.exception(
+                            "Heartbeat failed to start.  This could result in a zombie run."
+                        )
             except Exception as exc:
                 self.logger.exception(
                     "Heartbeat failed to start.  This could result in a zombie run."
                 )
             return runner_method(self, *args, **kwargs)
         finally:
-            timer.cancel()
+            was_alive = all([t.is_alive() for t in heartbeat_threads])
+            was_running = timer.cancel()
+            if not (was_alive and was_running):
+                self.logger.warning(
+                    "Heartbeat thread appears to have died.  This could result in a zombie run."
+                )
 
     return inner